replaced float

+++ src/self_e.C
@@ -1,7 +1,7 @@
 #include "libinclude.h"
 #include "const.h"
 
-double selfe(int n_atomtype, int *natoms_type, float *chargs, float betaa){
+double selfe(int n_atomtype, int *natoms_type, float *chargs, double betaa){
     double self_energy=0;
     for (int i = 0; i < n_atomtype; i++){
         self_energy+=natoms_type[i]*pow(chargs[i],2);
